Use merge ingredient table instead of sum
